refactored SearchService
